Cleanup contrib
